+++ include/linalg.tpp
@@ -13,7 +13,7 @@ Vector<K> linearCombination(std::vector<Vector<K> >& u, std::vector<K>& coefs) {
 		K sum = 0;
 		for (size_t index = 0; index < u.size(); ++index) {
 			const std::vector<K>& data = u[index].getData();
-			sum = std::fma(data[dim], coefs[index], sum);
+			sum = std::fmaf(data[dim], coefs[index], sum);
 		}
 		result.setCoord(dim, sum);
 	}
@@ -24,7 +24,7 @@ Vector<K> linearCombination(std::vector<Vector<K> >& u, std::vector<K>& coefs) {
 template <typename V>
 V lerp(V u, V v, float t) {
 	if constexpr (std::is_arithmetic_v<V>) {
-		return std::fma(v - u, t, u);
+		return std::fmaf(v - u, t, u);
 	}
 	return u + (v - u) * t;
 }
@@ -40,8 +40,8 @@ Vector<K> cross_product(Vector<K>& u, Vector<K>& v) {
 		throw std::invalid_argument("Both vector u and v should be 3D.");
 	}
 
-	K p0 = std::fma(u[1], v[2], -(std::fma(u[2], v[1], 0)));
-	K p1 = std::fma(u[2], v[0], -(std::fma(u[0], v[2], 0)));
-	K p2 = std::fma(u[0], v[1], -(std::fma(u[1], v[0], 0)));
+	K p0 = std::fmaf(u[1], v[2], -(std::fmaf(u[2], v[1], 0)));
+	K p1 = std::fmaf(u[2], v[0], -(std::fmaf(u[0], v[2], 0)));
+	K p2 = std::fmaf(u[0], v[1], -(std::fmaf(u[1], v[0], 0)));
 	return {p0, p1, p2};
 }
